# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/user":{"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"get":{"operationId":"UserController_getProfileInfo","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/auth/enable-2fa":{"post":{"operationId":"AuthController_enableTwoFa","parameters":[],"responses":{"201":{"description":""}}}},"/api/v1/auth/verify-2fa":{"post":{"operationId":"AuthController_verifyAuth","parameters":[],"responses":{"201":{"description":""}}}},"/api/v1/auth/refresh-token":{"get":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"@daco/auth APIs Specification","description":"Auth APIs description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"refreshToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"phone":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","phone","firstName","lastName"]},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"Aracely95@gmail.com"},"password":{"type":"string","example":"jTpQB4TuOGJq6Vt"}},"required":["email","password"]},"UserCreateDto":{"type":"object","properties":{"email":{"type":"string","example":"Baron.Kozey97@yahoo.com"},"password":{"type":"string","example":"04RfBK80063lz9Z"},"firstName":{"type":"string","example":"Adaline"},"lastName":{"type":"string","example":"Stehr"},"username":{"type":"string","example":"Roderick_Keebler98"}},"required":["email","password","firstName","lastName","username"]}}}}